temporarily let test fail for php 8

+++ plugins/CoreUpdater/tests/Integration/Commands/UpdateTest.php
@@ -102,6 +102,10 @@ class UpdateTest extends ConsoleCommandTestCase
 
     public function test_UpdateCommand_ReturnsCorrectExitCode_WhenErrorOccurs()
     {
+        if (PHP_MAJOR_VERSION == 8) {
+            $this->fail('Test terminates on PHP 8 for some reason, so temporarily directly let it fail');
+        }
+
         // create a blob table, then drop it manually so update 2.10.0-b10 will fail
         $tableName = ArchiveTableCreator::getBlobTable(Date::factory('2015-01-01'));
         Db::exec("DROP TABLE $tableName");
